tests: Multi project support
