+++ tests/test_integration/conftest.py
@@ -1,23 +1,10 @@
 import pytest
-import contextlib
 
-import os
 from cognite.powerops.clients.powerops_client import get_powerops_client
 from cognite.powerops.clients.powerops_client import PowerOpsClient
 from pathlib import Path
 from tests.constants import REPO_ROOT
-
-
-@contextlib.contextmanager
-def chdir(new_dir: Path) -> None:
-    current_working_dir = os.getcwd()
-    os.chdir(new_dir)
-
-    try:
-        yield
-
-    finally:
-        os.chdir(current_working_dir)
+from tests.utils import chdir
 
 
 @pytest.fixture(scope="session")
